% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul.R
\name{calcul_total_moyen_par_monument}
\alias{calcul_total_moyen_par_monument}
\title{Calculer la fréquentation moyenne par monument}
\usage{
calcul_total_moyen_par_monument(data)
}
\arguments{
\item{data}{Un data frame contenant les données de fréquentation.}
}
\value{
Un tibble avec le nom du monument et sa fréquentation moyenne.
}
\description{
Cette fonction calcule la fréquentation moyenne pour chaque monument.
}
