% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtre.R
\name{filtrer_monument}
\alias{filtrer_monument}
\title{Filtrer les données par monument}
\usage{
filtrer_monument(data, monument)
}
\arguments{
\item{data}{Un data frame contenant les données de fréquentation.}

\item{monument}{Le nom du monument à sélectionner.}
}
\value{
Un data frame filtré.
}
\description{
Cette fonction filtre le jeu de données selon le nom d'un monument.
}
