% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphique.R
\name{plot_total_moyen_par_monument}
\alias{plot_total_moyen_par_monument}
\title{Representation de la frequentation moyenne des monuments}
\usage{
plot_total_moyen_par_monument(data, top_n = 10)
}
\arguments{
\item{data}{Un data frame contenant les donnees.}

\item{top_n}{Le nombre de monuments a afficher.}
}
\value{
Un graphique ggplot2.
}
\description{
Cette fonction produit un graphique des monuments les plus frequentes.
}
